+++ src/index.ts
@@ -1,7 +1,7 @@
 export { parse } from './parse.js'
 export { parseStream } from './stream.js'
 
-export { CostLimitError } from './types.js'
+export { OpenParseError, InvalidDocumentError, UnsupportedFormatError, CostLimitError } from './types.js'
 
 export type {
   ParseOptions,